endpoint: fix incorrect semantics for is_same_host
The algorithm in CoAP 6.4.5 permits eliding Uri-Host only if the
hostname from the URI is an IP literal matching the endpoint.  DNS
lookup must not be done.

Also correct mis-use of getaddrinfo flags: AI_NUMERICSERV means the
service should not be looked up by name if it's already an integer; keep
that flag and verify that port is always an int.  AI_NUMERICHOST means
the node must already be in numeric format; this may not be the case
when endpoints are created, so remove the flag.