endpoint: initial message-layer behavior support
At this point we have "send" which creates a cache entry, cache entries
that manage BEBO and state changes related to reply reception, a
"receive" that reads data and records the cache entry.  There's some
initial handling of error cases, but not complete.

Checkpointing what's there so we can move on to evolving the user-level
API.  Expect all this to be significantly refactored after that.