endpoint: add statistics
To verify we are meeting congestion requirements we need to know when we
last heard from an endpoint, and how much we've sent it since we last
heard from it.  For analysis it's also good to know overall how much has
been sent to and received from the endpoint.  Support all this at the
level of raw send and receive.