endpoint: refactor statistics and remote state
Move message ID management down to LocalEndpoint, and support a normal
practice where the message ID is assigned when the message is
transmitted.  Also collect all statistics and state that are relevant to
a remote endpoint in an association class that links the source endpoint
to that remote endpoint, since the source endpoint is also relevant to
the state.